+++ API/Services/Catalog/Units/IUnitServices.cs
@@ -6,10 +6,10 @@ namespace API.Services.Catalog.Units
 {
     public interface IUnitServices
     {
-        Task<ApiResult<byte>> Create(UnitCreateRequest request);
-        Task<ApiResult<byte>> Update(UnitUpdateRequest request);
-        Task<ApiResult<byte>> Delete(byte unitId);
+        Task<ApiResult<int>> Create(UnitCreateRequest request);
+        Task<ApiResult<int>> Update(UnitUpdateRequest request);
+        Task<ApiResult<int>> Delete(int unitId);
         Task<ApiResult<PagedResult<UnitViewModel>>> GetPaging(UnitGetPagingRequest request);
-        Task<ApiResult<UnitViewModel>> GetById(byte unitId);
+        Task<ApiResult<UnitViewModel>> GetById(int unitId);
     }
 }
